+++ application/views/home/home.js
@@ -6,13 +6,14 @@ var divResultsArticles = $('#homeTopArticles');
 var divResultsRefreshTime = $('#lastRefreshTime');
 // Ajax call
 function getLatestArticles() {
- var go = $.ajax({
-     type: 'POST',
-     url: 'APIURL/articles/getArticles',
-     data: {
-         m: 'ajax'
-     }
- })
+  var params = { 'm': 'ajax'};
+  var go = $.ajax({
+    url: 'APIURL/articles/getArticles',
+    type: 'POST',
+    dataType: 'json',
+    contentType: 'application/json;chartset=utf-8',
+    data: JSON.stringify(params)
+  })
  .done(function(results) {
      var now = new Date();
      divResultsRefreshTime.html(now.toString());
@@ -20,7 +21,8 @@ function getLatestArticles() {
      divResultsArticles.css('border', '3px solid black');
  })
  .fail(function(msg) {
-     alert('Error: ' + msg);
+   divResultsRefreshTime.html('Oops!');
+   console.debug(msg);
  })
  .always(function() {
  });