+++ public/index.php
@@ -29,11 +29,9 @@ if (filter_has_var(INPUT_GET, 'url')) {
   $url = 'home';
 }
 
-/*
- * 
 // Domain info
-$domains = new DomainsController('Domains', 'Domain', 'getDomains');
-$domainInfo = $domains->lookupDomains(NULL, $domain)[0];
+$domains = new DomainsController('Domains', 'Domain', 'getDomains', 'json');
+$domainInfo = $domains->getDomains(NULL, $domain)[0];
 $deny = TRUE;
 if (!empty($domainInfo)) {
   if ((int)$domainInfo['domainActive'] === 1) {
@@ -51,8 +49,8 @@ if ($deny === TRUE) {
 }
 
 // Brand info
-$brands = new BrandsController('Brands', 'Brand', 'getBrands');
-$brandInfo = $brands->lookupBrands($brandID)[0];
+$brands = new BrandsController('Brands', 'Brand', 'getBrands', 'json');
+$brandInfo = $brands->getBrands($brandID)[0];
 $deny = TRUE;
 if (!empty($brandInfo)) {
   if ((int)$brandInfo['brandActive'] === 1) {
@@ -66,11 +64,6 @@ if ($deny === TRUE) {
   trigger_error('Brand not found: '.$brandID, E_USER_ERROR);
   die();
 }
- * 
- */
-define('DOMAINURL', 'http://'.$_SERVER['HTTP_HOST'].'/');
-define('DOMAIN', $domain);
-define('BRAND', $domain);
     
 // Debug Info
 if (ERROR_LEVEL == 10